refactor(webui): modularize file viewer components into registry-driven architecture
