+++ packages/webui/components/viewers/image/image-control-bar.tsx
