+++ packages/webui/components/viewers/video/progress-bar.tsx
